+++ objectbox-java/src/main/java/io/objectbox/sync/SyncClientImpl.java
@@ -16,6 +16,7 @@
 
 package io.objectbox.sync;
 
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
@@ -43,7 +44,7 @@ public final class SyncClientImpl implements SyncClient {
 
     @Nullable
     private BoxStore boxStore;
-    private final String serverUrl;
+    private final List<String> urls;
     private final InternalSyncClientListener internalListener;
     @Nullable
     private final ConnectivityMonitor connectivityMonitor;
@@ -62,13 +63,41 @@ public final class SyncClientImpl implements SyncClient {
 
     SyncClientImpl(SyncBuilder builder) {
         this.boxStore = builder.boxStore;
-        this.serverUrl = builder.serverUrl();
+        this.urls = builder.urls;
         this.connectivityMonitor = builder.platform.getConnectivityMonitor();
 
-        long boxStoreHandle = builder.boxStore.getNativeStore();
-        long handle = nativeCreate(boxStoreHandle, serverUrl, builder.trustedCertPaths);
+        // Build the options
+        long optHandle = nativeSyncOptCreate(builder.boxStore.getNativeStore());
+        if (optHandle == 0) {
+            throw new RuntimeException("Failed to create Sync client options: handle is zero.");
+        }
+        try {
+            // Add all server URLs
+            for (String url : urls) {
+                nativeSyncOptAddUrl(optHandle, url);
+            }
+
+            // Add trusted certificate paths if provided
+            if (builder.trustedCertPaths != null) {
+                for (String certPath : builder.trustedCertPaths) {
+                    nativeSyncOptAddCertPath(optHandle, certPath);
+                }
+            }
+
+            // Add Sync flags if set
+            if (builder.flags != 0) {
+                nativeSyncOptFlags(optHandle, builder.flags);
+            }
+        } catch (Exception e) {
+            // Free the options if any option method call failed (like due to invalid arguments)
+            nativeSyncOptFree(optHandle);
+            throw e;
+        }
+
+        // Create the sync client (this frees the options in any case)
+        long handle = nativeSyncOptCreateClient(optHandle);
         if (handle == 0) {
-            throw new RuntimeException("Failed to create sync client: handle is zero.");
+            throw new RuntimeException("Failed to create Sync client: handle is zero.");
         }
         this.handle = handle;
 
@@ -101,13 +130,7 @@ public final class SyncClientImpl implements SyncClient {
         this.internalListener = new InternalSyncClientListener();
         nativeSetListener(handle, internalListener);
 
-        if (builder.credentials.size() == 1) {
-            setLoginCredentials(builder.credentials.get(0));
-        } else if (builder.credentials.size() > 1) {
-            setLoginCredentials(builder.credentials.toArray(new SyncCredentials[0]));
-        } else {
-            throw new IllegalArgumentException("No credentials provided");
-        }
+        setLoginCredentials(builder.credentials);
 
         // If created successfully, let store keep a reference so the caller does not have to.
         InternalAccess.setSyncClient(builder.boxStore, this);
@@ -123,7 +146,13 @@ public final class SyncClientImpl implements SyncClient {
 
     @Override
     public String getServerUrl() {
-        return serverUrl;
+        // nativeSyncOptCreateClient guarantees there is at least one URL
+        return getUrls().get(0);
+    }
+
+    @Override
+    public List<String> getUrls() {
+        return urls;
     }
 
     @Override
@@ -202,6 +231,17 @@ public final class SyncClientImpl implements SyncClient {
 
     public void removeAllFilterVariables() {
         nativeRemoveAllFilterVariables(getHandle());
+    }
+
+    @Override
+    public void setLoginCredentials(List<SyncCredentials> credentials) {
+        if (credentials.size() == 1) {
+            setLoginCredentials(credentials.get(0));
+        } else if (credentials.size() > 1) {
+            setLoginCredentials(credentials.toArray(new SyncCredentials[0]));
+        } else {
+            throw new IllegalArgumentException("Credentials must be provided");
+        }
     }
 
     @Override
@@ -356,10 +396,61 @@ public final class SyncClientImpl implements SyncClient {
     }
 
     /**
-     * Creates a native sync client for the given store handle ready to connect to the server at the given URI.
-     * Uses certificate authorities trusted by the host if no trusted certificate paths are passed.
+     * Creates a sync client options object for the given store.
+     * <p>
+     * The options must be configured (at least one URL) and then used with {@link #nativeSyncOptCreateClient}.
+     *
+     * @return handle to the options object, or 0 on error
      */
-    private static native long nativeCreate(long storeHandle, String uri, @Nullable String[] certificateDirsOrPaths);
+    private static native long nativeSyncOptCreate(long storeHandle);
+
+    /**
+     * Adds a server URL to the sync options; at least one URL must be added before creating the sync client.
+     * <p>
+     * Passing multiple URLs allows high availability and load balancing (i.e. using an ObjectBox Sync Server Cluster).
+     * <p>
+     * A random URL is selected for each connection attempt.
+     */
+    private static native void nativeSyncOptAddUrl(long optHandle, String url);
+
+    /**
+     * Adds a certificate path to the sync options.
+     * <p>
+     * This allows to pass certificate paths referring to the local file system.
+     * <p>
+     * Example use cases are using self-signed certificates in a local development environment and custom certificate
+     * authorities.
+     */
+    private static native void nativeSyncOptAddCertPath(long optHandle, String certPath);
+
+    /**
+     * Sets sync flags to adjust sync behavior; see SyncFlags for available flags.
+     * <p>
+     * Combine multiple flags using bitwise OR.
+     */
+    private static native void nativeSyncOptFlags(long optHandle, int flags);
+
+    /**
+     * Creates a sync client with the given options.
+     * <p>
+     * This does not initiate any connection attempts yet: call {@link #nativeStart} to do so. Before nativeStart(), you
+     * must configure credentials via {@link #nativeSetLoginInfo} or {@link #nativeAddLoginCredentials}.
+     * <p>
+     * By default, a sync client automatically receives updates from the server once login succeeded. To configure this
+     * differently, call {@link #nativeSetRequestUpdatesMode} with the wanted mode.
+     * <p>
+     * Note: the given options are always freed by this function, including when an error occurs.
+     *
+     * @return handle to the sync client, or 0 on error
+     */
+    private static native long nativeSyncOptCreateClient(long optHandle);
+
+    /**
+     * Frees the sync options object.
+     * <p>
+     * Note: Only free *unused* options; {@link #nativeSyncOptCreateClient} frees the options internally.
+     */
+    private static native void nativeSyncOptFree(long optHandle);
 
     private native void nativeDelete(long handle);